Move viewmorepopup
